Reverting changes According to LM feedback
'Asincrónico' is the approved term for Microsoft so proposed change is not acceptable. 
Also note that 'Azure Functions' must be left in English if this is the service name. According to context found in translated file sent by email, this seems to be the service name, therefore the Spanish translation must be fixed accordingly.