informal word used
The word “quitar” (and its derivative for 3rd person singular “quita”) is kind of a vulgar word that we would never use in a formal situation as this one because it has a literal meaning of physically taking something out. The verb “Eliminar” and “elimina” are correct and appropriate for this translation.