feminine singular “específica” to complement VM
Specific is an adjective complementing VM, not to sizes (given that the articles talks about the sizes of a specific VM, not about the specific sizes of a VM), in Spanish we adjust the adjective to the name it is complementing to. In this case it should be feminine singular “específica”, not masculine plural.